+++ src/components/layout/Layout.tsx
@@ -4,6 +4,8 @@ import { Inter } from 'next/font/google'
 import Head from 'next/head'
 import Script from 'next/script'
 
+import AdSlot from '@/components/ui/AdSlot'
+
 const inter = Inter({
   subsets: ['latin'],
   variable: '--font-inter',
@@ -30,6 +32,7 @@ export default function Layout({ children }: { children: ReactNode }): ReactElem
         `}
       </Script>
 
+      <AdSlot />
       <main className={`${inter.className} h-full min-h-screen bg-base text-content`}>{children}</main>
     </>
   )