+++ src/components/settings/Settings.tsx
@@ -13,7 +13,6 @@ import GeneralSettings from '@/components/settings/GeneralSettings'
 import Logs from '@/components/settings/Logs'
 import TradingCardManagerSettings from '@/components/settings/TradingCardManagerSettings'
 import ViewDocumentation from '@/components/settings/ViewDocumentation'
-import AdSlot from '@/components/ui/AdSlot'
 import useSettings from '@/hooks/settings/useSettings'
 
 export default function Settings(): ReactElement {
@@ -39,8 +38,6 @@ export default function Settings(): ReactElement {
       </div>
 
       <div className='absolute flex flex-col items-center gap-4 bottom-4 left-0 px-6 w-[250px] z-40'>
-        {process.env.NODE_ENV === 'production' && <AdSlot />}
-
         <ViewDocumentation />
         <span className='text-xs text-altwhite text-center'>Steam Game Idler v{version}</span>
       </div>